# Micro-optimization to AmazingPrint::ActiveRecord#cast